requirements.txt
django==3.2
pillow
python-dotenv
djangorestframework
drf-yasg